# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "crucible",
    "description": "REST API for the crucible daemon, a Firecracker microVM sandbox runtime. The daemon is the contract every SDK mirrors. Auth is a bearer token (`Authorization: Bearer \u003ckey\u003e`); `/healthz` is always exempt, and a loopback daemon with no keys configured serves unauthenticated. The exec and service-log endpoints stream a binary frame protocol OpenAPI cannot model — see the guides for the frame format.",
    "version": "0.3.2"
  },
  "paths": {
...
    "/sandboxes/{id}/service": {
      "get": {
        "tags": [
          "service"
        ],
        "summary": "Service status",
        "operationId": "serviceStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireServiceStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "service"
        ],
        "summary": "Configure the supervised service",
        "description": "Experimental. Sets and (re)starts the guest's supervised long-lived entrypoint.",
        "operationId": "configureService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureServiceReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireServiceStatus"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{id}/service/logs": {
      "get": {
        "tags": [
          "service"
        ],
        "summary": "Stream service logs",
        "description": "Experimental. Streams the supervised service's stdout/stderr as it is produced.",
        "operationId": "serviceLogs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "base64"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{id}/service/restart": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Restart the service",
        "operationId": "restartService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireServiceStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{id}/service/start": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Start the service",
        "operationId": "startService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireServiceStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{id}/service/stop": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Stop the service",
        "operationId": "stopService",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Sandbox ID, e.g. sbx_ab12cd34.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Sandbox ID, e.g. sbx_ab12cd34."
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceStopReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireServiceStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConfigureServiceReq": {
        "type": "object",
        "properties": {
          "cmd": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "cwd": {
            "type": "string"
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "env_exact": {
            "type": "boolean"
          },
          "log_buffer_bytes": {
            "type": "integer"
          },
          "restart": {
            "$ref": "#/components/schemas/WireRestartPolicy"
          },
          "stop_grace_s": {
            "type": "integer"
          },
          "stop_signal": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ServiceStopReq": {
        "type": "object",
        "properties": {
          "grace_s": {
            "type": "integer"
          }
        }
      },
...
      "WireExecResult": {
        "type": "object",
        "properties": {
          "duration_ms": {
            "type": "integer",
            "format": "int64"
          },
          "error": {
            "type": "string"
          },
          "exit_code": {
            "type": "integer"
          },
          "oom_killed": {
            "type": "boolean"
          },
          "signal": {
            "type": "string"
          },
          "timed_out": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/WireResourceUsage"
          }
        }
      },
...
      "WireResourceUsage": {
        "type": "object",
        "properties": {
          "context_switches_involuntary": {
            "type": "integer",
            "format": "int64"
          },
          "cpu_sys_ms": {
            "type": "integer",
            "format": "int64"
          },
          "cpu_user_ms": {
            "type": "integer",
            "format": "int64"
          },
          "io_read_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "io_write_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "page_faults_major": {
            "type": "integer",
            "format": "int64"
          },
          "peak_memory_bytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "WireRestartPolicy": {
        "type": "object",
        "properties": {
          "max_retries": {
            "type": "integer"
          },
          "policy": {
            "type": "string"
          }
        }
      },
      "WireServiceSpec": {
        "type": "object",
        "properties": {
          "cmd": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "cwd": {
            "type": "string"
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "env_exact": {
            "type": "boolean"
          },
          "log_buffer_bytes": {
            "type": "integer"
          },
          "restart": {
            "$ref": "#/components/schemas/WireRestartPolicy"
          },
          "stop_grace_s": {
            "type": "integer"
          },
          "stop_signal": {
            "type": "string"
          },
          "user": {
            "type": "string"
          }
        }
      },
      "WireServiceStatus": {
        "type": "object",
        "properties": {
          "last_exit": {
            "$ref": "#/components/schemas/WireExecResult"
          },
          "last_exit_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "last_exit_requested": {
            "type": "boolean"
          },
          "live_peak_rss_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "live_rss_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "log_dropped_bytes": {
            "minimum": 0,
            "type": "integer"
          },
          "log_first_seq": {
            "minimum": 0,
            "type": "integer"
          },
          "log_next_seq": {
            "minimum": 0,
            "type": "integer"
          },
          "pid": {
            "type": "integer"
          },
          "restarts": {
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/WireServiceSpec"
          },
          "started_at_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "type": "string"
          },
          "uptime_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Daemon API key. Omit on a keyless loopback daemon."
      }
    }
  }
}